Refactoring - singleton generator
