+++ DesignPatternCodeGenerator/Singleton/SingletonGenerator.cs
@@ -2,7 +2,9 @@
 using DesignPatternCodeGenerator.Base.Enums;
 using DesignPatternCodeGenerator.Base.Generators;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
+using System.Linq;
 using System.Text;
 
 namespace DesignPatternCodeGenerator.Singleton
@@ -21,12 +23,18 @@ namespace DesignPatternCodeGenerator.Singleton
 
             foreach (var group in classGroups)
             {
-                var classContent = SingletonContentGenerator.GenerateClass(group);
-
-                context.AddSource(
-                    $"{BaseNamesGenerator.GetClassName(group, GeneratorAttributeType.Singleton).Replace("Singleton", "")}.g.cs",
-                    SourceText.From(classContent, Encoding.UTF8));
+                GenerateSingleton(context, group);
             }
+        }
+
+        private void GenerateSingleton(
+            GeneratorExecutionContext context,
+            IGrouping<string, ClassDeclarationSyntax> group)
+        {
+            var hintName = $"{BaseNamesGenerator.GetClassName(group)}.g.cs";
+            var classContent = SingletonContentGenerator.GenerateClass(group);
+
+            context.AddSource(hintName, SourceText.From(classContent, Encoding.UTF8));
         }
 
         public void Initialize(GeneratorInitializationContext context)
